+Finalized first tests
