--- src/testingHelper.ts
@@ -1,19 +1,0 @@
-import { MongooseModule, MongooseModuleOptions } from '@nestjs/mongoose';
-import { MongoMemoryServer } from 'mongodb-memory-server';
-
-let mongod: MongoMemoryServer;
-
-export const rootMongooseTestModule = (options: MongooseModuleOptions = {}) => MongooseModule.forRootAsync({
-  useFactory: async () => {
-    mongod = new MongoMemoryServer();
-    const mongoUri = await mongod.getUri();
-    return {
-      uri: mongoUri,
-      ...options,
-    }
-  },
-});
-
-export const closeInMongodConnection = async () => {
-  if (mongod) await mongod.stop();
-}
